requirement.txt
streamlit
pandas
mysql.connector